add more metadata
